Refactor swarm support
