+++ strelets/adapter/swarm.go
@@ -8,8 +8,6 @@ import (
 	"github.com/docker/docker/client"
 	"io"
 	"os"
-	"strings"
-	"time"
 )
 
 type dockerSwarm struct {
@@ -62,121 +60,6 @@ func (d *dockerSwarm) RemoveContainer(ctx context.Context, containerName string)
 	return d.client.ServiceRemove(ctx, getServiceId(containerName))
 }
 
-func (d *dockerSwarm) storeConfiguration(ctx context.Context, containerName string, root string, config *AppConfig) (*dockerSwarmSecretsConfig, error) {
-	secrets := &dockerSwarmSecretsConfig{}
-
-	if keyPairSecretId, err := d.saveSwarmSecret(ctx, containerName, "keyPair", config.KeyPair); err != nil {
-		return nil, fmt.Errorf("could not store key pair secret: %s", err)
-	} else {
-		secrets.keysSecretId = keyPairSecretId
-	}
-
-	if networkSecretId, err := d.saveSwarmSecret(ctx, containerName, "network", config.Network); err != nil {
-		return nil, fmt.Errorf("could not store network config secret: %s", err)
-	} else {
-		secrets.networkSecretId = networkSecretId
-	}
-
-	return secrets, nil
-}
-
-func (d *dockerSwarm) Prepare(ctx context.Context, imageName string, containerName string, root string, httpPort int, gossipPort int, appConfig *AppConfig) (Runner, error) {
-	config, err := d.storeConfiguration(ctx, containerName, root, appConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	ureplicas := uint64(1)
-	restartDelay := time.Duration(10 * time.Second)
-
-	keysSecret := &swarm.SecretReference{
-		SecretName: getSwarmSecretName(containerName, "keyPair"),
-		SecretID:   config.keysSecretId,
-		File: &swarm.SecretReferenceFileTarget{
-			Name: "keys.json",
-			UID:  "0",
-			GID:  "0",
-		},
-	}
-
-	networkSecret := &swarm.SecretReference{
-		SecretName: getSwarmSecretName(containerName, "network"),
-		SecretID:   config.networkSecretId,
-		File: &swarm.SecretReferenceFileTarget{
-			Name: "network.json",
-			UID:  "0",
-			GID:  "0",
-		},
-	}
-
-	spec := swarm.ServiceSpec{
-		TaskTemplate: swarm.TaskSpec{
-			ContainerSpec: &swarm.ContainerSpec{
-				Image: "orbs:export",
-				Command: []string{
-					"/opt/orbs/orbs-node",
-					"--silent",
-					"--config", "/var/run/secrets/keys.json",
-					"--config", "/var/run/secrets/network.json",
-					// FIXME add separate volume for logs
-					"--log", "/opt/orbs/node.log",
-				},
-				Secrets: []*swarm.SecretReference{
-					keysSecret,
-					networkSecret,
-				},
-			},
-			RestartPolicy: &swarm.RestartPolicy{
-				Delay: &restartDelay,
-			},
-		},
-		Mode: swarm.ServiceMode{
-			Replicated: &swarm.ReplicatedService{
-				Replicas: &ureplicas,
-			},
-		},
-		EndpointSpec: &swarm.EndpointSpec{
-			Ports: []swarm.PortConfig{
-				{
-					Protocol:      "tcp",
-					PublishMode:   swarm.PortConfigPublishModeIngress,
-					PublishedPort: uint32(httpPort),
-					TargetPort:    8080,
-				},
-				{
-					Protocol:      "tcp",
-					PublishMode:   swarm.PortConfigPublishModeIngress,
-					PublishedPort: uint32(gossipPort),
-					TargetPort:    4400,
-				},
-			},
-		},
-	}
-	spec.Name = getServiceId(containerName)
-
-	return &dockerSwarmRunner{
-		client: d.client,
-		spec:   spec,
-	}, nil
-}
-
 func getServiceId(input string) string {
 	return "stack-" + input
-}
-
-func (d *dockerSwarm) saveSwarmSecret(ctx context.Context, containerName string, secretName string, content []byte) (string, error) {
-	secretId := getSwarmSecretName(containerName, secretName)
-	d.client.SecretRemove(ctx, secretId)
-
-	secretSpec := swarm.SecretSpec{
-		Data: content,
-	}
-	secretSpec.Name = secretId
-
-	response, err := d.client.SecretCreate(ctx, secretSpec)
-	return response.ID, err
-}
-
-func getSwarmSecretName(containerName string, secretName string) string {
-	return strings.Join([]string{containerName, secretName}, "-")
 }
